multiple minor changes
- added header row for DGA and  Crypto / Cryptojacking tables
- added C2Hunter IoCs sources into the analysis report
- displaying C2Hunter indicators in the analysis report only if the plugin is enabled
- rounded duration to the nearest whole number in long connections

+++ src/detection_engine.py
@@ -209,7 +209,7 @@ class DetectionEngine:
                             src_port=src_port,
                             dst_ip=dst_ip,
                             dst_port=dst_port,
-                            duration=float(duration)
+                            duration=int(duration)
                         )
                         detected_connections.append(entry)
                 else:
